moving test to it's own file due to env var overwriting
